refactor(audio): move write implementation to file layer
